Update ProgressSteps component
