+++ cardstack/src/components/ProgressSteps/index.ts
@@ -1,2 +1,3 @@
 export * from './ProgressSteps';
 export * from './CustomScrollView';
+export * from './StepIcon';
